requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pillow